+++ dt-eval-cli/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## [0.1.3-alpha](https://github.com/dynatrace-oss/dt-evals/compare/v0.1.2-alpha...v0.1.3-alpha) (2026-04-30)
+
+
+### 🐛 Bug Fixes
+
+* fix build for CLI ([#54](https://github.com/dynatrace-oss/dt-evals/issues/54)) ([591b20e](https://github.com/dynatrace-oss/dt-evals/commit/591b20eb53236d96804f0081a8df07bfc922fa4e))
+* **main:** redact secretKey in configure --show output ([#51](https://github.com/dynatrace-oss/dt-evals/issues/51)) ([a77ca6c](https://github.com/dynatrace-oss/dt-evals/commit/a77ca6c44e682df55d3f41b7ae3f3ccc04729b0d))
+
+## [0.1.2-alpha](https://github.com/dynatrace-oss/dt-evals/compare/v0.1.1-alpha...v0.1.2-alpha) (2026-04-28)
+
+
+### ✨ New Features
+
+* add Azure OpenAI and Bedrock provider support ([#18](https://github.com/dynatrace-oss/dt-evals/issues/18)) ([0988a0e](https://github.com/dynatrace-oss/dt-evals/commit/0988a0e73ebd6db17e7eacfdf3f08b5830665ba4))
+* **main:** prompt to run eval immediately after setup wizard ([#46](https://github.com/dynatrace-oss/dt-evals/issues/46)) ([9377f21](https://github.com/dynatrace-oss/dt-evals/commit/9377f218e4e86ddc7aa033f54a372a03b810971d))
+
+
+### 🐛 Bug Fixes
+
+* **dt-eval-lib:** remove Node.js builtins via PromptStore DI ([#43](https://github.com/dynatrace-oss/dt-evals/issues/43)) ([b6e8f39](https://github.com/dynatrace-oss/dt-evals/commit/b6e8f393d09b932a9cbcd12f3dc546ead416bf1d))
+
+
+### 📚 Documentation
+
+* update README for bedrock, azure-openai, and optional sampling ([#42](https://github.com/dynatrace-oss/dt-evals/issues/42)) ([f879ea4](https://github.com/dynatrace-oss/dt-evals/commit/f879ea46ffa30018190efa642b6c4f7a7718e255))
+
 ## [0.1.1-alpha](https://github.com/dynatrace-oss/dt-evals/compare/v0.1.0-alpha...v0.1.1-alpha) (2026-04-17)